refactor: make dashboard component composable
